add heroes routes test
 - cleanup main.py routes to live within the heroes router
 - add tests for heroes routes
