+++ src/tests/test_main.py
@@ -7,11 +7,3 @@ def test_get_root():
     response = client.get('/')
     assert response.status_code == 200
     assert response.json() == { 'message': 'Hello World'}
-
-def test_create_hero(client: TestClient):
-    response = client.post(
-        '/heroes/', 
-        json={'name': 'Deadpond', 'secret_name': 'Dive Wilson'}
-    )
-
-    assert response.status_code == 200
